requirements.txt
django
gunicorn
python-dotenv
whitenoise
django-import-export
pillow
git+https://github.com/JoseLArantes/django-admin-soft-dashboard.git


# psycopg2-binary
# mysqlclient
